training/requirements.txt
ultralytics==8.3.34
opencv-python-headless==4.10.0.84
numpy
matplotlib
pyyaml
tqdm
pillow
pdf2image
labelme
